+++ CHANGES.md
@@ -21,6 +21,7 @@ Internals:
 Bug fixes:
 - Correctly calculate the bounding box for primitives with scaled coordinates and radius bigger than the bounding box (#60)
 - Fix cones lighting (#65)
+- Scale axes ticks position (#74)
 
 1.1.0
 -----